[refactor] #98 다이어리 선택 시 다음 화면으로 안넘어가는 문제 해결
